+++ include/dlaf/matrix.h
@@ -16,10 +16,11 @@
 #include "dlaf/matrix/internal/tile_future_manager.h"
 #include "dlaf/matrix/layout_info.h"
 #include "dlaf/matrix/matrix_base.h"
-#include "dlaf/tile.h"
+#include "dlaf/matrix/tile.h"
 #include "dlaf/types.h"
 
 namespace dlaf {
+namespace matrix {
 
 /// A @c Matrix object represents a collection of tiles which contain all the elements of a matrix.
 ///
@@ -112,8 +113,6 @@ private:
   using Matrix<const T, device>::tile_managers_;
 };
 
-#include "dlaf/matrix.tpp"
-
 template <class T, Device device>
 class Matrix<const T, device> : public matrix::internal::MatrixBase {
 public:
@@ -166,8 +165,6 @@ private:
 
   std::vector<matrix::internal::TileFutureManager<T, device>> tile_managers_;
 };
-
-#include "dlaf/matrix_const.tpp"
 
 // Note: the templates of the following helper functions are inverted w.r.t. the Matrix templates
 // to allow the user to only specify the device and let the compiler deduce the type T.
@@ -338,3 +335,9 @@ Matrix<T, device> createMatrixFromTile(const GlobalElementSize& size, const Tile
 }
 
 }
+/// Make dlaf::matrix::Matrix available as dlaf::Matrix.
+using matrix::Matrix;
+}
+
+#include "dlaf/matrix.tpp"
+#include "dlaf/matrix_const.tpp"
