+++ include/dlaf/matrix/tile.h
@@ -28,13 +28,19 @@ struct ContinuationException final : public std::runtime_error {
       : std::runtime_error("An exception has been thrown during the execution of the previous task.") {}
 };
 
+namespace matrix {
+namespace internal {
+template <class T, Device device>
+class ViewTileFutureManager;
+}
+
 template <class T, Device device>
 class Tile;
 
 template <class T, Device device>
 class Tile<const T, device>;
 
-/// @brief The Tile object aims to provide an effective way to access the memory as a two dimensional
+/// The Tile object aims to provide an effective way to access the memory as a two dimensional
 /// array. It does not allocate any memory, but it references the memory given by a @c MemoryView object.
 /// It represents the building block of the Matrix object and of linear algebra algorithms.
 ///
@@ -47,11 +53,13 @@ class Tile<const T, device>;
 template <class T, Device device>
 class Tile<const T, device> {
   friend Tile<T, device>;
+  friend internal::ViewTileFutureManager<T, device>;
 
 public:
   using ElementType = T;
 
-  /// @brief Constructs a (@p size.rows() x @p size.cols()) Tile.
+  /// Constructs a (@p size.rows() x @p size.cols()) Tile.
+  ///
   /// @throw std::invalid_argument if @p size.row() < 0, @p size.cols() < 0 or @p ld < max(1, @p size.rows()).
   /// @throw std::invalid_argument if memory_view does not contain enough elements.
   /// The (i, j)-th element of the Tile is stored in the (i+ld*j)-th element of memory_view.
@@ -61,7 +69,7 @@ public:
 
   Tile(Tile&& rhs) noexcept;
 
-  /// @brief Destroys the Tile.
+  /// Destroys the Tile.
   /// If a promise was set using @c setPromise its value is set to a Tile
   /// which has the same size and which references the same memory as @p *this.
   ~Tile();
@@ -70,7 +78,8 @@ public:
 
   Tile& operator=(Tile&& rhs) noexcept;
 
-  /// @brief Returns the (i, j)-th element,
+  /// Returns the (i, j)-th element,
+  ///
   /// where @p i := @p index.row and @p j := @p index.col.
   /// @pre index.isValid() == true.
   /// @pre index.isIn(size()) == true.
@@ -78,12 +87,13 @@ public:
     return *ptr(index);
   }
 
-  /// @brief Returns the base pointer.
+  /// Returns the base pointer.
   const T* ptr() const noexcept {
     return memory_view_();
   }
 
-  /// @brief Returns the pointer to the (i, j)-th element,
+  /// Returns the pointer to the (i, j)-th element,
+  ///
   /// where @p i := @p index.row and @p j := @p index.col.
   /// @pre index.isValid() == true.
   /// @pre index.isIn(size()) == true.
@@ -96,17 +106,17 @@ public:
     return memory_view_(sum(index.row(), mul(ld_, index.col())));
   }
 
-  /// @brief Returns the size of the Tile.
+  /// Returns the size of the Tile.
   const TileElementSize& size() const noexcept {
     return size_;
   }
-  /// @brief Returns the leading dimension.
+  /// Returns the leading dimension.
   SizeType ld() const noexcept {
     return ld_;
   }
 
 private:
-  /// @brief Sets size to {0, 0} and ld to 1.
+  /// Sets size to {0, 0} and ld to 1.
   void setDefaultSizes() noexcept;
 
   TileElementSize size_;
@@ -123,7 +133,8 @@ class Tile : public Tile<const T, device> {
 public:
   using ElementType = T;
 
-  /// @brief Constructs a (@p size.rows() x @p size.cols()) Tile.
+  /// Constructs a (@p size.rows() x @p size.cols()) Tile.
+  ///
   /// @throw std::invalid_argument if @p size.row() < 0, @p size.cols() < 0 or @p ld < max(1, @p size.rows()).
   /// @throw std::invalid_argument if memory_view does not contain enough elements.
   /// The (i, j)-th element of the Tile is stored in the (i+ld*j)-th element of memory_view.
@@ -138,7 +149,8 @@ public:
 
   Tile& operator=(Tile&& rhs) = default;
 
-  /// @brief Returns the (i, j)-th element,
+  /// Returns the (i, j)-th element,
+  ///
   /// where @p i := @p index.row and @p j := @p index.col.
   /// @pre index.isValid() == true.
   /// @pre index.isIn(size()) == true.
@@ -146,12 +158,13 @@ public:
     return *ptr(index);
   }
 
-  /// @brief Returns the base pointer.
+  /// Returns the base pointer.
   T* ptr() const noexcept {
     return memory_view_();
   }
 
-  /// @brief Returns the pointer to the (i, j)-th element,
+  /// Returns the pointer to the (i, j)-th element,
+  ///
   /// where @p i := @p index.row and @p j := @p index.col.
   /// @pre index.isValid() == true.
   /// @pre index.isIn(size()) == true.
@@ -164,7 +177,8 @@ public:
     return memory_view_(sum(index.row(), mul(ld_, index.col())));
   }
 
-  /// @brief Sets the promise to which this Tile will be moved on destruction.
+  /// Sets the promise to which this Tile will be moved on destruction.
+  ///
   /// @c setPromise can be called only once per object.
   /// @throw std::logic_error if @c setPromise was already called.
   Tile& setPromise(hpx::promise<Tile<T, device>>&& p) {
@@ -181,13 +195,16 @@ private:
   using Tile<const T, device>::p_;
 };
 
-/// @brief Create a common::Buffer from a Tile
+/// Create a common::Buffer from a Tile
 template <class T, Device device>
 auto create_buffer(const Tile<T, device>& tile) {
   return common::Buffer<T*>(tile.ptr({0, 0}), to_sizet(tile.size().cols()), to_sizet(tile.size().rows()),
                             to_sizet(tile.ld()));
 }
 
-#include <dlaf/tile.tpp>
-
 }
+// Make dlaf::matrix::Tile available as dlaf::Tile.
+using matrix::Tile;
+}
+
+#include <dlaf/matrix/tile.tpp>
