+++ include/dlaf/matrix/tile.tpp
@@ -8,6 +8,9 @@
 // SPDX-License-Identifier: BSD-3-Clause
 //
 
+namespace dlaf {
+namespace matrix {
+
 template <class T, Device device>
 Tile<const T, device>::Tile(const TileElementSize& size,
                             memory::MemoryView<ElementType, device>&& memory_view, SizeType ld)
@@ -56,4 +59,7 @@ template <class T, Device device>
 void Tile<const T, device>::setDefaultSizes() noexcept {
   size_ = {0, 0};
   ld_ = 1;
+}
+
+}
 }
