+++ test/unit/CMakeLists.txt
@@ -12,9 +12,7 @@ DLAF_addTest(test_util_math SOURCES test_util_math.cpp USE_MAIN PLAIN)
 
 DLAF_addTest(test_blas_tile SOURCES test_blas_tile.cpp USE_MAIN HPX)
 DLAF_addTest(test_lapack_tile SOURCES test_lapack_tile.cpp USE_MAIN HPX)
-DLAF_addTest(test_tile SOURCES test_tile.cpp USE_MAIN HPX)
 
-DLAF_addTest(test_matrix SOURCES test_matrix.cpp USE_MAIN MPIHPX MPIRANKS 6)
 
 # Generic libraries
 add_subdirectory(common)
