+++ test/unit/communication/test_broadcast_tile.cpp
@@ -12,11 +12,11 @@
 
 #include <gtest/gtest.h>
 
+#include "dlaf/matrix/tile.h"
+
 #include "dlaf_test/helper_communicators.h"
 #include "dlaf_test/matrix/util_tile.h"
 #include "dlaf_test/util_types.h"
-
-#include "dlaf/tile.h"
 
 using namespace dlaf;
 using namespace dlaf::comm;
