+++ test/unit/matrix/CMakeLists.txt
@@ -8,8 +8,13 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 
+DLAF_addTest(test_tile SOURCES test_tile.cpp USE_MAIN HPX)
+
 DLAF_addTest(test_layout_info SOURCES test_layout_info.cpp USE_MAIN PLAIN)
-DLAF_addTest(test_util_distribution SOURCES test_util_distribution.cpp USE_MAIN PLAIN)
 DLAF_addTest(test_distribution SOURCES test_distribution.cpp USE_MAIN PLAIN)
+DLAF_addTest(test_util_distribution SOURCES test_util_distribution.cpp USE_MAIN PLAIN)
+
+DLAF_addTest(test_matrix SOURCES test_matrix.cpp USE_MAIN MPIHPX MPIRANKS 6)
 DLAF_addTest(test_matrix_view SOURCES test_matrix_view.cpp USE_MAIN MPIHPX MPIRANKS 6)
+
 DLAF_addTest(test_util_matrix SOURCES test_util_matrix.cpp USE_MAIN MPIHPX MPIRANKS 6)
