+++ test/unit/test_blas_tile/test_gemm.h
@@ -14,8 +14,8 @@
 #include <sstream>
 #include "gtest/gtest.h"
 #include "dlaf/blas_tile.h"
+#include "dlaf/matrix/tile.h"
 #include "dlaf/memory/memory_view.h"
-#include "dlaf/tile.h"
 #include "dlaf/util_blas.h"
 #include "dlaf_test/matrix/util_tile.h"
 #include "dlaf_test/matrix/util_tile_blas.h"
