+++ test/unit/test_lapack_tile/test_potrf.h
@@ -13,8 +13,8 @@
 #include <sstream>
 #include "gtest/gtest.h"
 #include "dlaf/lapack_tile.h"
+#include "dlaf/matrix/tile.h"
 #include "dlaf/memory/memory_view.h"
-#include "dlaf/tile.h"
 #include "dlaf/util_blas.h"
 #include "dlaf_test/matrix/util_tile.h"
 #include "dlaf_test/util_types.h"
